Extract locators and steps for sharing in Files app to their own context
